Extract shared node creation commands
To new superclass 'CreateNodeCommand'
The commonalities are the selction of the node name, location and the
reading/creation of the node (Now in `run` which calls a class specific
`action`) and the tempfile mangement (extracted to edit_with_tmp_file)
