+++ lib/commands/create_node_command.rb
@@ -20,15 +20,36 @@
 # https://github.com/alces-software/inventoryware
 #==============================================================================
 
+require 'tempfile'
+require 'tty-editor'
+
 module Inventoryware
   module Commands
-    class Edit < CreateNodeCommand
-      def action(node_data, location)
-        # output to create the node's file if it doesn't yet exist
-        Utils::output_node_yaml(node_data, location)
-        # maybe don't create unless saved? i.e. don't create the file above
-        # instead save as closing
-        TTY::Editor.open(location, command: :rvim)
+    class CreateNodeCommand < Command
+      def run
+        name = @argv[0]
+        location = File.join(YAML_DIR, "#{name}.yaml")
+        node_data = Utils::read_node_or_create(location)
+
+        action(node_data, location)
+      end
+
+      def action
+        raise NotImplementedError
+      end
+
+      def edit_with_tmp_file(text, command)
+        tmp_file = Tempfile.new('inv_ware_file_')
+        begin
+          TTY::Editor.open(tmp_file.path,
+                           content: text,
+                           command: command)
+          edited = tmp_file.read
+        ensure
+          tmp_file.close
+          tmp_file.unlink
+        end
+        return edited
       end
     end
   end
